Add tests for the new features in rhproxy 1.5.5.
This commit checks the Cloud Console for systems tagged to use Insights
proxy, removes the workarounds for previous rhc issues. In addition,
it's now possible to check the image version. Lastly, a bug in the
handling of the custom rhproxy RPM has been fixed. The README has been
updated accordingly.